unified pipeline and extras tab
I take lpw_stable_diffusion_onnx.py from the diffusers package and call it directly to use as a megapipeline. Also added SD1.5 inpainting, so it should work with both legacy and new inpainting models.
This pipeline uses torch.Generator for random seeds. In addition to making previous seeds incompatible (but in exchange for being seed-compatible with other UIs such as A1111), this may or may not hurt performance, but them's the breaks.
VAE encoder and decoder now are unloaded from memory after use, and are reloaded upon clicking Generate. This helps to keep them out of the way of UNET, instead of just sitting there wasting VRAM.
Moved deepdanbooru to it's own tab, allowing for the creation of more extra  features without cluttering up UI.